authorization added

+++ routers/posts.py
@@ -8,6 +8,7 @@ from sqlalchemy.orm import selectinload
 import models
 from database import get_db
 from schemas import PostCreate, PostResponse, PostUpdate, PostUpdatePartial
+from auth import CurrentUser
 
 router = APIRouter()
 
@@ -24,19 +25,16 @@ async def get_posts(db: AsyncSession = Depends(get_db)):
         response_model=PostResponse,
         status_code=status.HTTP_201_CREATED
 )
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-    result = await db.execute(select(models.User).where(models.User.id == post.user_id))
-    user = result.scalars().first()
-    if not user:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="User not found",
-        )
+async def create_post(
+    post: PostCreate, 
+    db: Annotated[AsyncSession, Depends(get_db)],
+    current_user: CurrentUser
+    ):
     
     post = models.Post(
         title=post.title,
         content=post.content,
-        user_id=post.user_id,
+        user_id=current_user.id,
     )
 
     db.add(post)
@@ -61,13 +59,23 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
 
 # update post fully
 @router.put('/{post_id}', response_model=PostResponse)
-async def update_post(post_id: int, post_data: PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post(
+    post_id: int, 
+    post_data: PostUpdate, 
+    db: Annotated[AsyncSession, Depends(get_db)],
+    current_user: CurrentUser
+    ):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not found",
+        )
+    if current_user.id != post.user_id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not authorized to update this post",
         )
     post.title = post_data.title
     post.content = post_data.content
@@ -77,13 +85,23 @@ async def update_post(post_id: int, post_data: PostUpdate, db: Annotated[AsyncSe
 
 # update post partially
 @router.patch('/{post_id}', response_model=PostResponse)
-async def update_post_partial(post_id: int, post_data: PostUpdatePartial, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partial(
+    post_id: int, 
+    post_data: PostUpdatePartial, 
+    db: Annotated[AsyncSession, Depends(get_db)],
+    current_user: CurrentUser
+    ):
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not found",
+        )
+    if current_user.id != post.user_id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not authorized to update this post",
         )
     update_data = post_data.model_dump(exclude_unset=True)
     for key, value in update_data.items():
@@ -94,13 +112,24 @@ async def update_post_partial(post_id: int, post_data: PostUpdatePartial, db: An
     return post
 
 @router.delete('/{post_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def delete_post(
+    post_id: int, 
+    db: Annotated[AsyncSession, Depends(get_db)],
+    current_user: CurrentUser
+    ):
+
     result = await db.execute(select(models.Post).where(models.Post.id == post_id))
     post = result.scalars().first()
     if not post:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Post not found",
+        )
+    
+    if current_user.id != post.user_id:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail="You are not authorized to delete this post",
         )
     await db.delete(post)
     await db.commit()
